project/backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
openai
langchain
langchain-text-splitters
langchain-community
langchain[google-genai]
langchain[openai]
langchain-huggingface
sentence-transformers
langchain-google-genai
langchain-core
beautifulsoup4
google-cloud-firestore
pytest
pytest-asyncio
httpx
pinecone
python-pptx